+++ bin/inline_forms_installer_core.rb
@@ -70,6 +70,12 @@ if ENV['using_sqlite'] == 'true'
     pool: 5
     timeout: 5000
 
+  test:
+    adapter: sqlite3
+    database: db/test.sqlite3
+    pool: 5
+    timeout: 5000
+
   END_DATABASEYML
 else
   create_file "config/database.yml", <<-END_DATABASEYML.strip_heredoc